pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-crisis_room"
version = "0.1.0"
description = "Crisis Room environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [



    "openenv-core[core]>=0.2.2",








]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]


server = "crisis_room.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["crisis_room", "crisis_room.server"]
package-dir = { "crisis_room" = ".", "crisis_room.server" = "server" }